+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
 ## v1.0.x
+* Added: Bluetooth: Show signal strenght of BMS in log by @mr-manuel
 * Added: Exclude a device from beeing used by the dbus-serialbattery driver by @mr-manuel
 * Added: Implement callback function for update by @seidler2547
-* Changed: Improved Jkbms_Ble driver by @seidler2547 & @mr-manuel
+* Added: JKBMS BLE - Show last five characters from the MAC address in the custom name (which is displayed in the device list) by @mr-manuel
+* Changed: Fixed error in `reinstall-local.sh` script for Bluetooth installation by @mr-manuel
 * Changed: Fixed typo in `config.ini` sample by @hoschult
+* Changed: Improved Jkbms_Ble driver by @seidler2547 & @mr-manuel
 
 ## v1.0.20230531